Refactored Panels.js and Panels.css:
Panlels.js: theme names are now being passed by props to the Panel builder and belong to the specific panel they were added to.
Panels.css: removed the theme names, and have the styling applied to the theme class name only, altered the animation of the text. When hovering over the panel, text zooms out with a box around it for better visibilty.